rename updates folder
